🚚 Extract internal mapper into its own file
